se agrega funcionalidad de calcular provisiones

+++ database/migrations/2022_02_09_113850_create_provisions_table.php
@@ -11,16 +11,16 @@ class CreateProvisionsTable extends Migration
         Schema::create('provisions', function (Blueprint $table) {
             $table->id();
             $table->foreignId('period_id')->references('id')->on('periods');
-            $table->integer('pension_contribution')->nullable();
-            $table->integer('occupational_risk_contributions')->nullable();
-            $table->integer('health_contribution')->nullable();
-            $table->integer('compensation_funds')->nullable();
-            $table->integer('layoffs')->nullable();
-            $table->integer('Interest_on_severance_pay')->nullable();
-            $table->integer('vacation_provision')->nullable();
-            $table->integer('wage_premium')->nullable();
-            $table->integer('total_provisions')->nullable();
-            $table->integer('total_payroll')->nullable();
+            $table->double('pension_contribution')->nullable();
+            $table->double('occupational_risk_contributions')->nullable();
+            $table->double('health_contribution')->nullable();
+            $table->double('compensation_funds')->nullable();
+            $table->double('layoffs')->nullable();
+            $table->double('Interest_on_severance_pay')->nullable();
+            $table->double('vacation_provision')->nullable();
+            $table->double('wage_premium')->nullable();
+            $table->double('total_provisions')->nullable();
+            $table->double('total_payroll')->nullable();
             $table->timestamps();
         });
     }
